git-svn-id: https://emgucv.svn.sourceforge.net/svnroot/emgucv/trunk@417 d7f09016-e345-0410-b530-edf29a71df78

+++ Emgu.CV.ML/ANN_MLP.cs
@@ -44,7 +44,7 @@ namespace Emgu.CV.ML
          Matrix<float> trainData,
          Matrix<float> responses,
          Matrix<float> sampleWeights,
-         Matrix<int> sampleIdx,
+         Matrix<Byte> sampleIdx,
          MCvANN_MLP_TrainParams parameters,
          MlEnum.ANN_MLP_TRAINING_FLAG flag)
       {
